refactor: Phase D - extract 5 composables from large components
Extract heavy logic into composables, leaving components as thin UI shells:

- useCoinSearchChat: streaming, provider status, image cache, save (~20 functions)
- useImageProcessor: crop math, canvas drawing, bg removal, save pipeline (~15 functions)
- useCollectionFilters: search, sort, category, tag, page sync (~8 functions)
- useSettingsProfile: avatar, profile CRUD, password, NumisBids (~10 functions)
- useAdminConfig: AI settings, test connections, model lists (~12 functions)

~1,034 lines removed from parent components. vue-tsc passes clean.
